pyproject.toml
[project]
name = "automaton-auditor"
version = "0.1.0"
description = "Hierarchical LangGraph agent swarm for autonomous code auditing"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langsmith>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "docling>=2.0.0",
    "pymupdf>=1.24.0",
    "Pillow>=10.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
